tests/requirements.txt
-r ../requirements.txt
pytest-django
model-bakery
